refactor(chat): split index into useChatSession and memoized message UI
Extract chat-types, chat-message-utils, use-chat-session, MessageStatsBar,
InitialThreadScrollInitializer, ChatMessageList, and memoized ChatMessageRow.
Shrink index.tsx to Chat composition only.
